Major overhaul completed. Missing generics in order to get rid of typecasting the cells and models in the closures
